Improvements

+++ mkctf/cli/command/delete.py
@@ -4,6 +4,9 @@
 async def delete(api, args):
     '''Deletes a challenge
     '''
+    if not args.yes and confirm(f"do you really want to run delete?") == Answer.NO:
+        app_log.warning("operation cancelled by user.")
+        return False
     result = api.delete(args.slug)
     return result['deleted']
